+++ HISTORY.rst
@@ -4,6 +4,11 @@ Version history
 0.3
 ---
 
+* ``--keep-history`` defaults to blank, which uses the original value
+  set in portal_purgehistory.  Now when the original value is -1
+  (unlimited) you can set it to 0 with this option.  Previously this
+  was ignored.
+
 * Added ``--permanent`` option.
 
 * Be more careful about keeping the history.